adds initial oauth device flow, groundwork for polling for access_token

+++ cmd/run.go
@@ -136,34 +136,83 @@ func publicKey(keyname string) (*ssh.PublicKeys, error) {
 }
 
 func run() error {
-	// Create a tempDir to clone into
-	tempDir, err := createTempDir()
+	var dryrun bool = true
 
-	if err != nil {
-		return fmt.Errorf("cannot create temporary directory: %s", err)
+	if !dryrun {
+		// Create a tempDir to clone into
+		tempDir, err := createTempDir()
+
+		if err != nil {
+			return fmt.Errorf("cannot create temporary directory: %s", err)
+		}
+
+		// Cleanup tempDir
+		defer os.Remove(tempDir)
+
+		// Clone the remote
+		repo, err := cloneRepo(repositoryURL, tempDir)
+
+		if err != nil {
+			return fmt.Errorf("cannot clone repository: %s", err)
+		}
+
+		// Create the tag
+		err = createTag(repo)
+		if err != nil {
+			return fmt.Errorf("cannot create tag: %s", err)
+		}
+
+		// Run a build
+		err = makeBuild(tempDir)
+		if err != nil {
+			return fmt.Errorf("failed building artifacts: %s", err)
+		}
 	}
 
-	// Cleanup tempDir
-	defer os.Remove(tempDir)
+	fmt.Println("did nothing; got here")
+	fmt.Printf(clientID)
 
-	// Clone the remote
-	repo, err := cloneRepo(repositoryURL, tempDir)
-
+	// Create a release
+	// AUTH:
+	//		request user & device codes
+	var scope string = ""
+	authResponse, err := requestDeviceAndUserCodes(githubEndpoint.DeviceAuthURL, clientID, scope)
 	if err != nil {
-		return fmt.Errorf("cannot clone repository: %s", err)
+		return fmt.Errorf("failed requesting device and user codes from github: %s", err)
 	}
+	fmt.Println(authResponse.DeviceCode)
+	fmt.Println(authResponse.UserCode)
+	fmt.Println(authResponse.VerificationURI)
+	fmt.Println(authResponse.ExpiresIn)
+	fmt.Println(authResponse.Interval)
 
-	// Create the tag
-	err = createTag(repo)
+	//		prompt user
+	fmt.Printf("Please enter your one-time verification code at %s\n", authResponse.VerificationURI)
+	fmt.Printf("One-time code: %s\n", authResponse.UserCode)
+	openbrowser(authResponse.VerificationURI)
+
+	userAuthResponse, err := pollForAccessToken(
+		githubEndpoint.TokenURL,
+		clientID,
+		authResponse.DeviceCode,
+		githubDeviceGrantType,
+		authResponse.ExpiresIn,
+		authResponse.Interval,
+	)
 	if err != nil {
-		return fmt.Errorf("cannot create tag: %s", err)
+		return fmt.Errorf("failed checking for authorization and retrieving access token: %s", err)
 	}
+	fmt.Println(userAuthResponse.AccessToken)
+	fmt.Println(userAuthResponse.TokenType)
+	fmt.Println(userAuthResponse.Scope)
 
-	// Run a build
-	err = makeBuild(tempDir)
-	if err != nil {
-		return fmt.Errorf("failed building artifacts: %s", err)
-	}
-
+	// Start timeout to Expires; check Interval ; move on when authed
+	//		poll for auth status
+	// List releases (does one exist?)
+	// https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#list-releases
+	// Create a Release
+	// https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#create-a-release
+	// Upload Release Assets
+	// https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#upload-a-release-asset
 	return nil
 }
